Chess.Board

Description

The state of the current game.

Documentation

initialBoard :: Board

The initial position of a chess game.

Properties

prop_FEN :: Board -> Bool

Property that asserts that writing a board in fen notaion and then reading it gives the same board back.