Chess.Board.Attacks

Description

Checks for a Square being attacked by one of the players.

https://chessprogramming.wikispaces.com/Square+Attacked+By

Documentation

isAttacked :: Board -> Colour -> Square -> Bool

are any of the given player's pieces attacking the given square?

Boolean test with sliding piece occupancy testing (magic look up).

attackedFromBB :: Board -> BitBoard -> Colour -> Square -> BitBoard

Bitboard with the position of the attacking pieces set. Occupancy can be specified, ie. we can remove pieces from the board.

inCheck :: Board -> Colour -> Bool

is the specified player in check?

inCheckWithNoFriendly :: Board -> Colour -> Bool

is the specified player in check with the friendly pieces removed?