Chess.Board.Board

Description

Data.BitBoard representation of the current state of the game

In search it's also used for storing nodes. It should support fast move and unmove operations. To query for instance the white pawns one can do

 (b^.whitePieces) .&. (b^.pawns)

Documentation

data Board

Constructors

emptyBoard :: Board

Utilities

opponent' :: Colour -> Colour

black for white, white for black

prettyPrint :: Board -> IO ()

fen :: Board -> String

Board lenses

whitePieces :: Lens' Board BitBoard

blackPieces :: Lens' Board BitBoard

rooks :: Lens' Board BitBoard

knights :: Lens' Board BitBoard

bishops :: Lens' Board BitBoard

queens :: Lens' Board BitBoard

kings :: Lens' Board BitBoard

pawns :: Lens' Board BitBoard

next :: Lens' Board Colour

opponent :: Lens' Board Colour

opposite colour of the next lens

enPassant :: Lens' Board (Maybe Square)

whiteCastleRights :: Lens' Board CastlingRights

blackCastleRights :: Lens' Board CastlingRights

hash :: Lens' Board Word64

Lenses by type

piecesByColour

Arguments

:: Colour
Black / White

-> Lens' Board BitBoard
Lens

the BitBoard Lens corresponding to the given colour

piecesByType

Arguments

:: PieceType
Rook / Pawn etc.

-> Lens' Board BitBoard
Lens

the BitBoard Lens corresponding to the given PieceType

castleRightsByColour :: Colour -> Lens' Board CastlingRights

Castle rights lens corresponding to the given colour

Queries

pieceAt :: Board -> Square -> Maybe PieceType

The piece type at the given position

pieceColourAt :: Board -> Square -> Maybe Colour

The piece colour at a given position

occupancy :: Board -> BitBoard

the occupancy Data.BitBoard\

vacated :: Board -> BitBoard

the empty squares Data.BitBoard\

myPieces :: Board -> BitBoard

my pieces

opponentsPieces :: Board -> BitBoard

opponents pieces

piecesOf :: Board -> Colour -> PieceType -> BitBoard

pieces of a player of a specific type

myPiecesOf :: Board -> PieceType -> BitBoard

opponentsPiecesOf :: Board -> PieceType -> BitBoard

numberOf :: Board -> Colour -> PieceType -> Int

Others

calcHash :: Board -> Word64

calculates hash from scratch