Chess.History

Description

Relative history heuristics

This is a relative history heuristics with butterfly penalties. The normal history heuristics are described here: https://chessprogramming.wikispaces.com/History+Heuristic. The relative penalty improvement can be found here: https://chessprogramming.wikispaces.com/Relative+History+Heuristic.

Documentation

data History

History store

Constructor

mkHistory :: History

The empty Store

Manipulation

insert :: Int -> Move -> [PseudoLegalMove] -> History -> History

Insert move at given depth (should be called on the beta cut). Also updates the buterfly penalties on the moves in the given move list up to the move in the list.

Query

heuristics

Arguments

:: [PseudoLegalMove]
previous move list

-> History

-> [PseudoLegalMove]

the new move list with the heuristics applied