Chess.Killer

Description

Killer move heuristics

https://chessprogramming.wikispaces.com/Killer+Heuristic. This is called from the search iteration loop with updates, where a fail high node would cause the update. The store is queried before the iteration and moves are rearranged accordingly.

Documentation

data Killer

at a given depth entries are sorted by hit count in reverse order

Constructor

mkKiller :: Killer

Creates a killer store

Data manipulation

insert

Arguments

:: Int
distance from start depth ( or in other words real depth )

-> Move
move to insert

-> Killer

-> Killer

inserts a Move to the store

clearLevel :: Int -> Killer -> Killer

Clears the specified depth of the store

Query

heuristics

Arguments

:: Int
distance from start depth ( or in other words real depth )

-> [PseudoLegalMove]
previous move list

-> Killer

-> [PseudoLegalMove]

the new move list with the heuristics applied