Chess.Magic

Description

The magic bit board database

The database helps quickly determining in run time the sliding pieces (Rook, Bishop) attack BitBoards including blocking piece checks. The total run time is a 64 entry lookup (per piece type) obtaining the magic dataset. From there we have the magicNumber which then can be used (involving 64-bit multiplication) combined with the current board occupancy BitBoard to obtain the magicIndex. This gives the desired BitBoard with the attacked positions set (including capturable pieces). The returned position does not distinguish between colours, therefore one might need to and the resulting BitBoard with the negate of the friendly pieces.

The per piece span varies on the 64 entry position, for rooks the maximal required span address width is 12 for any position. (12 is maximum of attackable non-edge squares with a rook). 1 entry is 64 bits, therefore the maximal database size is 8 bytes * (2 ^ 12 span width) * 64 entries. This is 2Mb however due to the fancy-BitBoard packing (spans are packed with no gaps) it is believed that we can go down to ~800Kb. For bishops the database is significantly smaller.

See:

the StockFish source code https://github.com/mcostalba/Stockfish

http://www.rivalchess.com/magic-bitboards

http://chessprogramming.wikispaces.com/Magic+Bitboards

Documentation

data Magic

The magic database ( for a Bishop or a Rook )

magic :: PieceType -> Square -> BitBoard -> BitBoard

properties

prop_slidingAttack :: BitBoard -> Square -> Property