Chess.Move.Generator

Legality checks

data PseudoLegalMove

A move that might leave our king in check

data Legality

mkPseudo :: Move -> PseudoLegalMove

Conversion from Move to PseudoLegalMove

mkLegality :: Board -> Legality

Move legality checker constructor

legalCheck :: Legality -> PseudoLegalMove -> (Legality, Maybe Move)

if the move is legal then Just move and the new Legality

capturedPiece' :: PseudoLegalMove -> Maybe PieceType

from' :: PseudoLegalMove -> Square

to' :: PseudoLegalMove -> Square

Generators

moves :: Board -> [PseudoLegalMove]

Pseudo legal moves

forcingMoves :: Board -> [PseudoLegalMove]

Checks, captures and promotions, or moving out of check

anyMove :: Board -> Bool

Is there a legal move?