Chess.Search.SearchState

Documentation

mkSearchState

Arguments

:: TVar Bool
the signal to toggle to abort a search

-> TVar Bool
the signal to toggle pondering

-> SearchState

Creates a search state with the initialBoard. Use the board Lens to manipulate the position in the SearchState

The TVar parameters can be changed asynchronously while the Search is running. The ponder should be set to false by the interface in response to ponderhit. The abort should be raised to kill a running Search and should be polled for the bit to be dropped as a signal back mechanism that the abort succeeded.

data SearchState

The state of the Search monad

Lenses

board :: Lens' SearchState Board

aborted :: Lens' SearchState (TVar Bool)

pondering :: Lens' SearchState (TVar Bool)

timeStats :: Lens' SearchState TimeStats

tpc :: Lens' SearchState TransPosCache

kill :: Lens' SearchState Killer

pv :: Lens' SearchState PVStore

tpcHit :: Lens' SearchState Int

tpcMiss :: Lens' SearchState Int

nCnt :: Lens' SearchState Int

clock :: Lens' SearchState (Maybe UTCTime)