Chess.TimeControl

Description

The module calculates the allocated time for the search based on the times remaining and other factors.

https://chessprogramming.wikispaces.com/Time+Management

The allocated time for the search is independent of the engine speed, and only depends on the remaining times, the moves played, and the state of the game. The search is supposed to call the timeControl function before entering each level of iterative deepening, to decide whether to proceed or terminate.

We can terminate the search prematurely if the move to be played is "obvious" ie the search is stable. For only legal moves or mate combinations the Search should avoid unnecessary calculations regardless of time control. Move stability depends on the oscillation of the score between the deepening iterations.

Documentation

data TimeControl

The time control setup as received from the UCI.

Constructors

Timed

Fields

whiteTime :: Time

White player's time

whiteInc :: Maybe Time

White player's increment

blackTime :: Time

Black player's time

blackInc :: Maybe Time

Black player's increment

movesToGo :: Maybe Int

Moves until time controll

DepthSpecified Int
Ignore time, use this depth

TimeSpecified Int
Run for this amount of time

Infinite
Run for ever

data TimeStats

The store for the search statistics. Storing the time taken, the evaluation, and the best move per depth level.

mkTimeStats :: TimeStats

Creates an empty TimeStats instance

addStats :: Time -> Maybe (Move, Int) -> TimeStats -> TimeStats

Adds the next level to the statistics

timeControl :: Colour -> Bool -> Int -> TimeControl -> TimeStats -> Bool

Time control. Based on pondering, the UCI time control, the number of remaining pieces, and our current time statistics decides whether we can proceed to the next level.

Setters

setWhiteTime :: Time -> TimeControl -> TimeControl

Sets white's time

setBlackTime :: Time -> TimeControl -> TimeControl

Sets black's time

setWhiteInc :: Time -> TimeControl -> TimeControl

Sets white's increment

setBlackInc :: Time -> TimeControl -> TimeControl

Sets black's increment

setMovesToGo :: Int -> TimeControl -> TimeControl

Sets the moves to go until the next time control