Chess.TransPosCache

Description

Transpositional cache

We return hit if all of the 3 conditions are met:

hash of Board matches

Board equals

depth is equal or less than the one stored in cache

The are no entries with only the depth differing. Therefore inserting the same position with different depth updates the previous entry. Also EXACT entry wins over LOWER or UPPER. For more info see http://chessprogramming.wikispaces.com/Transposition+Table.

Documentation

type TransPosCache = LRU Word64 TransPosCacheEntry

data TransPosCacheEntry

data TransPosCacheEntryType

Constructors

Exact Move [Move]
An exact lookup always contains at least one move.

Lower Move
A lower bound cut contains just one move.

Upper

Constructors

mkTransPosCache :: TransPosCache

Lenses

board :: Lens' TransPosCacheEntry Board

depth :: Lens' TransPosCacheEntry Int

score :: Lens' TransPosCacheEntry Int

typ :: Lens' TransPosCacheEntry TransPosCacheEntryType

Utils

transPosCacheLookUp

Arguments

:: Board

-> Int
depth

-> TransPosCache

-> Either (Maybe Move) (TransPosCache, TransPosCacheEntry)

Either a move recommendation or Nothing on miss or the cache entry with the updated LRU on hit

transPosCacheInsert

Arguments

:: Board
board

-> Int
depth

-> Int
score

-> TransPosCacheEntryType
type

-> TransPosCache

-> TransPosCache

Returns a cache with the entry inserted

transPosCacheDeflate :: TransPosCache -> TransPosCache

decreases the depth of each entry by 1