Chess.Zobrist

Description

The module provides a zobrist keys from a vector of 781 pseudo random bitboards.

This can be used for a positions zobrist hash calculations. For more info see http:/chessprogramming.wikispaces.comZobrist+Hashing.

Documentation

data ZobristKey

Constructors

ZobristPiece

Fields

square :: Square

pieceColor :: Colour

pieceType :: PieceType

ZobristSide

Fields

side :: Colour

ZobristCastlingRights

Fields

whiteRights :: CastlingRights

blackRights :: CastlingRights

ZobristEnPassant

Fields

enPassantSquare :: Maybe Square

zobrist :: ZobristKey -> Word64