Data.BitBoard.TH

Description

Template haskell functions

Documentation

toVecLookup

Arguments

:: (Enum a, Bounded a)

=> String
name of the new function

-> (a -> BitBoard)
Enum to BitBoard function

-> Q [Dec]
Resulting splice

Transforms a function that maps a Bounded Enum to BitBoards from pattern matching (or whatever way the function is implemented) to a vector lookup.

toVecLookup2

Arguments

:: (Enum a, Bounded a, Enum b, Bounded b)

=> String
Name of the function

-> Integer
TODO : remove this somehow - the maxBound + 1 of the 2nd enum

-> (a -> b -> BitBoard)
the tranformed function

-> Q [Dec]
resulting new splice

Transforms a function that maps two Bounded Enums to BitBoards from pattern matching (or whatever way the function is implemented) to a vector lookup.