Data.Square

Types

data Square

The square represents one of the squares of a chess board

data File

A File on a Board

data Rank

A rank on a chess board

All elements of a given type domain

squares :: [Square]

files :: [File]

ranks :: [Rank]

Files

aFile :: File

bFile :: File

cFile :: File

dFile :: File

eFile :: File

fFile :: File

gFile :: File

hFile :: File

Ranks

firstRank :: Rank

secondRank :: Rank

thirdRank :: Rank

fourthRank :: Rank

fifthRank :: Rank

sixthRank :: Rank

seventhRank :: Rank

eighthRank :: Rank

Conversions

file :: Square -> File

Extracts the file from a square

rank :: Square -> Rank

Extracts the Rank from a square

toSquare :: File -> Rank -> Square

Constructs a square from a file and a rank

Distances

hDiff :: Square -> Square -> Int

Horizonatal difference (can be negative)

hDist :: Square -> Square -> Int

Horizonatal distance

vDiff :: Square -> Square -> Int

Vertical difference (can be negative)

vDist :: Square -> Square -> Int

Vertical distance

offset :: Square -> Int -> Square

Offset a square by an Int (in the enumeration of a1 .. h1 .. a8 .. h8

utilities

mirror :: Square -> Square

Horizontal miror. Mirrors on the line between 4th and 5th ranks

parserSquare :: Parser Square

The parser for a1 etc notation