Chess.Board.Attacks

Documentation

attackedBy :: PieceType -> Board -> BitBoard -> Colour -> Square -> BitBoard

the bitboard from where the piece type of the given colour is attacking the specified position

isAttacked :: Board -> Colour -> Square -> Bool

are any of the given player's pieces attacking the given square?

inCheck :: Board -> Colour -> Bool

inCheckWithNoFriendly :: Board -> Colour -> Bool

is the king with the given colour attacked with the friendly pieces removed?