Chess.Board.Board

Description

Representation of the current state of the game

Documentation

data Board

Constructors

emptyBoard :: Board

A board with no pieces

Data manipulation

flipPiece :: Colour -> PieceType -> Either Square BitBoard -> Board -> Board

For Kings a Left Square should be given and this simply sets the King position. For other piece types a Right BitBoard flips the presence of the specified piece type using XOR logic.

removePiece :: Square -> Board -> Board

Removes a piece (except the king) from a square

Lenses

next :: Lens' Board Colour

opponent :: Lens' Board Colour

opposite colour of the next lens

enPassant :: Lens' Board (Maybe Square)

whiteCastleRights :: Lens' Board CastlingRights

blackCastleRights :: Lens' Board CastlingRights

hash :: Lens' Board Word64

castleRightsByColour :: Colour -> Lens' Board CastlingRights

Castle rights lens corresponding to the given colour

Queries

pawns :: Board -> BitBoard

BitBoard with all the Pawns

piecesByColour :: Board -> Colour -> BitBoard

the occupancy bitboard for a given colour

pieceAt :: Board -> Square -> Maybe PieceType

The piece type at the given position

pieceColourAt :: Board -> Square -> Maybe Colour

The piece colour at a given position

occupancy :: Board -> BitBoard

the occupancy Data.BitBoard\

vacated :: Board -> BitBoard

the empty squares Data.BitBoard\

myPieces :: Board -> BitBoard

my pieces

opponentsPieces :: Board -> BitBoard

opponents pieces

piecesOf :: Board -> Colour -> PieceType -> BitBoard

pieces of a player of a specific type

kingByColour :: Colour -> Board -> Square

the king's square for a given colour

myPiecesOf :: Board -> PieceType -> BitBoard

opponentsPiecesOf :: Board -> PieceType -> BitBoard

numberOf :: Board -> Colour -> PieceType -> Int

Utilities

prettyPrint :: Board -> IO ()

fen :: Board -> String

The FEN string (without the halfmove, fullmove counters)

calcHash :: Board -> Word64

calculates hash from scratch

QuickCheck

prop_Board :: Board -> Bool

Property that asserts that the pieces are consistent. The union of black and white pieces is the same as the union of all piece types.