Chess.Killer

Description

Killer move heuristics

https:chessprogramming.wikispaces.com/Killer+Heuristic. This is called from the search iteration loop with updates, where a fail high node would cause the update. The store is queried before the iteration and moves are rearranged accordingly. This should be avoided on quiscene searches, as a killer move usually supposed to be some quiet move.

Now it's also used to store PV between different depth of iterative depening.

Documentation

data Killer

Constructor

mkKiller :: Killer

Data manipulation

insertKiller

Arguments

:: Int
distance from start depth ( or in other words real depth )

-> Maybe Move

-> Killer

-> Killer

insertPVInKiller :: [Move] -> Killer

killer store from the PV

Query

killer

Arguments

:: Int
distance from start depth ( or in other words real depth )

-> [Move]
previous move list

-> Killer

-> [Move]